requirements.txt
google
asgiref==3.2.10
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dj-config-url==0.1.1
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
django-filters==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
drf-jwt==1.16.1
idna==2.9
jsonfield==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==3.2
gunicorn==20.0.4
pyjwt==1.7.1
social-auth-app-django==3.1
azure-ai-textanalytics